requirements.txt
numpy==1.14.3
Pillow==6.2.2
streamlit==0.56.0